feat : create TravelTransaction entity
