+++ src/main/java/com/ajou/travely/controller/cost/dto/CostCreateRequestDto.java
@@ -1,12 +1,10 @@
 package com.ajou.travely.controller.cost.dto;
 
-import com.ajou.travely.domain.Cost;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
-import java.util.HashMap;
 import java.util.Map;
 
 @Getter
