+++ src/main/java/com/ajou/travely/controller/cost/dto/CostResponseDto.java
@@ -1,8 +1,7 @@
 package com.ajou.travely.controller.cost.dto;
 
 import com.ajou.travely.controller.user.dto.SimpleUserInfoDto;
-import com.ajou.travely.domain.Cost;
-import com.ajou.travely.domain.UserCost;
+import com.ajou.travely.domain.cost.Cost;
 import lombok.Getter;
 
 import java.util.List;
