+++ src/main/java/com/ajou/travely/repository/CostRepository.java
@@ -1,6 +1,6 @@
 package com.ajou.travely.repository;
 
-import com.ajou.travely.domain.Cost;
+import com.ajou.travely.domain.cost.Cost;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
