+++ src/main/java/com/ajou/travely/service/CostService.java
@@ -3,7 +3,7 @@ package com.ajou.travely.service;
 import com.ajou.travely.controller.cost.dto.CostCreateResponseDto;
 import com.ajou.travely.controller.cost.dto.CostResponseDto;
 import com.ajou.travely.controller.cost.dto.*;
-import com.ajou.travely.domain.Cost;
+import com.ajou.travely.domain.cost.Cost;
 import com.ajou.travely.domain.travel.Travel;
 import com.ajou.travely.domain.UserCost;
 import com.ajou.travely.exception.ErrorCode;
