+++ src/test/java/com/ajou/travely/service/CostServiceTest.java
@@ -2,7 +2,7 @@ package com.ajou.travely.service;
 
 import com.ajou.travely.controller.cost.dto.*;
 import com.ajou.travely.controller.travel.dto.TravelCreateRequestDto;
-import com.ajou.travely.domain.Cost;
+import com.ajou.travely.domain.cost.Cost;
 import com.ajou.travely.domain.UserCost;
 import com.ajou.travely.domain.travel.Travel;
 import com.ajou.travely.domain.user.User;
